pyproject.toml
[project]
name = "tatneft-test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.3",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "mypy>=1.19.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "ruff>=0.15.6",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true
